island: Move IslandConfigOptions validation to separate module
